+++ api/users.php
@@ -1,5 +1,7 @@
 <?php
-session_start();
+if (session_status() == PHP_SESSION_NONE) {
+    session_start();
+}
 
 require_once( 'siteadmins.php' );
 
@@ -10,6 +12,8 @@ require_once( 'siteadmins.php' );
 // getUser
 $app->get( '/users', function () {
     checkAuthorization( Slim::getInstance()->request() );
+    checkForAdmin();
+
 
     if( isset( $_GET[ 'admin' ] ) && $_GET[ 'admin' ] == 1 ) {
         $query = 'SELECT id, user_email FROM users WHERE admin == 1;';
@@ -28,6 +32,8 @@ $app->get( '/users', function () {
 // get user info
 $app->get( '/users/:id', function ( $user_id ) {
     checkAuthorization( Slim::getInstance()->request() );
+    checkForAdmin();
+
     $query = 'SELECT user_email FROM users WHERE id = :user_id;';
 
     try {
@@ -51,6 +57,8 @@ $app->get( '/users/:id', function ( $user_id ) {
 $app->delete( '/users/:id', function ( $user_id ) {
     $request = Slim::getInstance()->request();
     checkAuthorization( $request );
+    checkForAdmin();
+
 
     $query = 'DELETE FROM users WHERE id = :user_id AND admin != 1;';
     try {
@@ -74,6 +82,8 @@ $app->delete( '/users/:id', function ( $user_id ) {
 $app->post( '/users/:id/sites', function ( $user_id ) {
     $request = Slim::getInstance()->request();
     checkAuthorization( $request );
+    checkForAdmin();
+
     $request_body = json_decode( $request->getBody() );
 
     if( !$request_body->sites ) {
@@ -110,6 +120,8 @@ $app->post( '/users/:id/sites', function ( $user_id ) {
 $app->delete( '/users/:user_id/sites/:language/:site_id', function ( $user_id, $language, $site_id ) {
     $request = Slim::getInstance()->request();
     checkAuthorization( $request );
+    checkForAdmin();
+
 
     $query = "DELETE FROM site_admins WHERE user_id = :user_id AND site_id = :site_id AND language = :language;";
     try {
@@ -126,6 +138,8 @@ $app->delete( '/users/:user_id/sites/:language/:site_id', function ( $user_id, $
 $app->put( '/users', function () {
     $request = Slim::getInstance()->request();
     checkAuthorization( $request );
+    checkForAdmin();
+
     $user = json_decode( $request->getBody() );
 
     $query = "UPDATE users SET user_email=:email";
@@ -141,6 +155,8 @@ $app->put( '/users', function () {
 $app->post( '/users', function () {
     $request = Slim::getInstance()->request();
     checkAuthorization( $request );
+    checkForAdmin();
+
     $user = json_decode( $request->getBody() );
 
     $query = 'INSERT INTO users ( user_email, admin) VALUES ( :user_email, :admin );';
